requirements.txt
tensorflow  # or tensorflow-gpu if you're using GPU acceleration
numpy
gym-chess 
python-chess
ipywidgets
IPython
